implement mui MarvelAppBar, attach it to app, refactoring
