+++ src/index.tsx
@@ -1,10 +1,16 @@
 import ReactDOM from 'react-dom/client';
-import Router from './container/Router';
+import MarvelAppBar from './components/MarvelAppBar/MarvelAppBar';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
+import Router from './router/Router';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
-root.render(<Router />);
+root.render(
+  <div>
+    <MarvelAppBar />
+    <Router />
+  </div>
+);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
